requirements.txt
duckduckgo-search
fastapi
uvicorn
python-dotenv
agno
openai
psycopg2-binary
pgvector
sqlalchemy
pydantic
PyPDF2
python-multipart